Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    double int x,y;
+    long int x,y;
     char op;
     scanf("%d %d %c",&x,&y,&op);
     switch (op){
